[IMP] repair_stock: use a non conflicting field for auxiliar transfers
`repair_id` is reserved to the stock.moves directly related
to the execution of the repair order and has associated logic
in odoo core code (like assigning the picking type)
To avoid unexpected behaviors, we use `related_repair_id` to
signal auxiliar stock moves that are not directly related to
repair execution.

To not break existing modules depending on it, support for
`repair_id` is kept but it is recommended to remove in v18.